+++ sdk/ai/azure-ai-projects/samples/agents/sample_agents_azure_ai_search.py
@@ -4,6 +4,7 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
+
 """
 DESCRIPTION:
     This sample demonstrates how to use agent operations with the 
@@ -34,7 +35,7 @@ USAGE:
 import os
 from azure.ai.projects import AIProjectClient
 from azure.identity import DefaultAzureCredential
-from azure.ai.projects.models import AzureAISearchTool, ConnectionType
+from azure.ai.projects.models import AzureAISearchTool, ConnectionType, ListSortOrder, MessageRole
 
 project_client = AIProjectClient.from_connection_string(
     credential=DefaultAzureCredential(),
@@ -52,7 +53,7 @@ for conn in conn_list:
 print(conn_id)
 
 # Initialize agent AI search tool and add the search index connection id
-ai_search = AzureAISearchTool(index_connection_id=conn_id, index_name="myindexname")
+ai_search = AzureAISearchTool(index_connection_id=conn_id, index_name="sample_index")
 
 # Create agent with AI search tool and process assistant run
 with project_client:
@@ -74,7 +75,7 @@ with project_client:
     message = project_client.agents.create_message(
         thread_id=thread.id,
         role="user",
-        content="What inventory is available currently?",
+        content="What is the temperature rating of the cozynights sleeping bag?",
     )
     print(f"Created message, ID: {message.id}")
 
@@ -108,6 +109,21 @@ with project_client:
     project_client.agents.delete_agent(agent.id)
     print("Deleted agent")
 
+    # [START populate_references_agent_with_azure_ai_search_tool]
     # Fetch and log all messages
-    messages = project_client.agents.list_messages(thread_id=thread.id)
-    print(f"Messages: {messages}")
+    messages = project_client.agents.list_messages(thread_id=thread.id, order=ListSortOrder.ASCENDING)
+    for message in messages.data:
+        if message.role == MessageRole.AGENT and message.url_citation_annotations:
+            placeholder_annotations = {
+                annotation.text: f" [see {annotation.url_citation.title}] ({annotation.url_citation.url})"
+                for annotation in message.url_citation_annotations
+            }
+            for message_text in message.text_messages:
+                message_str = message_text.text.value
+                for k, v in placeholder_annotations.items():
+                    message_str = message_str.replace(k, v)
+                print(f"{message.role}: {message_str}")
+        else:
+            for message_text in message.text_messages:
+                print(f"{message.role}: {message_text.text.value}")
+    # [END populate_references_agent_with_azure_ai_search_tool]